+++ hw/components/tg/tg_device.h
@@ -33,18 +33,18 @@ public:
      *   Obtained from father
      *   void send_rsp (bool bErr);
      */
-    virtual void rcv_rqst (unsigned long ofs, unsigned char be,
+    virtual void rcv_rqst (unsigned int ofs, unsigned char be,
                            unsigned char *data, bool bWrite);
 
 private:
-    void write (unsigned long ofs, unsigned char be, unsigned char *data, bool &bErr);
-    void read  (unsigned long ofs, unsigned char be, unsigned char *data, bool &bErr);
+    void write (unsigned int ofs, unsigned char be, unsigned char *data, bool &bErr);
+    void read  (unsigned int ofs, unsigned char be, unsigned char *data, bool &bErr);
 
     void reset_input ();
 
 private:
     FILE                *tg_file;
-    unsigned long       tg_bytes_left;
+    unsigned int        tg_bytes_left;
 };
 
 #endif
